+++ tests/test_delayed_queue.py
@@ -17,14 +17,13 @@ from __future__ import annotations
 from time import time
 
 import pytest
-
 from watchdog.utils.delayed_queue import DelayedQueue
 
 
 @pytest.mark.flaky(max_runs=5, min_passes=1)
 def test_delayed_get():
     q = DelayedQueue[str](2)
-    q.put("", True)
+    q.put("", delay=True)
     inserted = time()
     q.get()
     elapsed = time() - inserted
@@ -35,7 +34,7 @@ def test_delayed_get():
 @pytest.mark.flaky(max_runs=5, min_passes=1)
 def test_nondelayed_get():
     q = DelayedQueue[str](2)
-    q.put("", False)
+    q.put("")
     inserted = time()
     q.get()
     elapsed = time() - inserted