+++ tests/test_fsevents.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
-import pytest
+import contextlib
 
+import pytest
 from watchdog.utils import platform
 
 if not platform.is_darwin():
@@ -14,36 +15,36 @@ from os import mkdir, rmdir
 from random import random
 from threading import Thread
 from time import sleep
+from typing import TYPE_CHECKING
 from unittest.mock import patch
 
 import _watchdog_fsevents as _fsevents  # type: ignore[import-not-found]
-
 from watchdog.events import FileSystemEventHandler
 from watchdog.observers import Observer
 from watchdog.observers.api import BaseObserver, ObservedWatch
 from watchdog.observers.fsevents import FSEventsEmitter
 
 from .shell import touch
-from .utils import P, StartWatching, TestEventQueue
+
+if TYPE_CHECKING:
+    from .utils import P, StartWatching, TestEventQueue
 
 logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)
 
 
-@pytest.fixture
+@pytest.fixture()
 def observer():
     obs = Observer()
     obs.start()
     yield obs
     obs.stop()
-    try:
+    with contextlib.suppress(RuntimeError):
         obs.join()
-    except RuntimeError:
-        pass
 
 
 @pytest.mark.parametrize(
-    "event,expectation",
+    ("event", "expectation"),
     [
         # invalid flags
         (_fsevents.NativeEvent("", 0, 0, 0), False),
@@ -119,7 +120,7 @@ def test_watcher_deletion_while_receiving_events_1(
         emitter = start_watching(tmpdir)
         # Less than 100 is not enough events to trigger the error
         for n in range(100):
-            touch(p("{}.txt".format(n)))
+            touch(p(f"{n}.txt"))
         emitter.stop()
         assert not caplog.records