+++ tests/test_observers_api.py
@@ -19,20 +19,19 @@ import time
 from pathlib import Path
 
 import pytest
-
 from watchdog.events import FileModifiedEvent, FileOpenedEvent, LoggingEventHandler
 from watchdog.observers.api import BaseObserver, EventDispatcher, EventEmitter, EventQueue, ObservedWatch
 
 
 def test_observer_constructor():
-    ObservedWatch(Path("/foobar"), True)
+    ObservedWatch(Path("/foobar"), recursive=True)
 
 
 def test_observer__eq__():
-    watch1 = ObservedWatch("/foobar", True)
-    watch2 = ObservedWatch("/foobar", True)
-    watch_ne1 = ObservedWatch("/foo", True)
-    watch_ne2 = ObservedWatch("/foobar", False)
+    watch1 = ObservedWatch("/foobar", recursive=True)
+    watch2 = ObservedWatch("/foobar", recursive=True)
+    watch_ne1 = ObservedWatch("/foo", recursive=True)
+    watch_ne2 = ObservedWatch("/foobar", recursive=False)
 
     assert watch1 == watch2
     assert watch1.__eq__(watch2)
@@ -41,10 +40,10 @@ def test_observer__eq__():
 
 
 def test_observer__ne__():
-    watch1 = ObservedWatch("/foobar", True)
-    watch2 = ObservedWatch("/foobar", True)
-    watch_ne1 = ObservedWatch("/foo", True)
-    watch_ne2 = ObservedWatch("/foobar", False)
+    watch1 = ObservedWatch("/foobar", recursive=True)
+    watch2 = ObservedWatch("/foobar", recursive=True)
+    watch_ne1 = ObservedWatch("/foo", recursive=True)
+    watch_ne2 = ObservedWatch("/foobar", recursive=False)
 
     assert not watch1.__ne__(watch2)
     assert watch1.__ne__(watch_ne1)
@@ -52,12 +51,12 @@ def test_observer__ne__():
 
 
 def test_observer__repr__():
-    observed_watch = ObservedWatch("/foobar", True)
+    observed_watch = ObservedWatch("/foobar", recursive=True)
     repr_str = "<ObservedWatch: path='/foobar', is_recursive=True>"
     assert observed_watch.__repr__() == repr(observed_watch)
     assert repr(observed_watch) == repr_str
 
-    observed_watch = ObservedWatch("/foobar", False, [FileOpenedEvent, FileModifiedEvent])
+    observed_watch = ObservedWatch("/foobar", recursive=False, event_filter=[FileOpenedEvent, FileModifiedEvent])
     repr_str = "<ObservedWatch: path='/foobar', is_recursive=False, event_filter=FileModifiedEvent|FileOpenedEvent>"
     assert observed_watch.__repr__() == repr(observed_watch)
     assert repr(observed_watch) == repr_str
@@ -65,14 +64,14 @@ def test_observer__repr__():
 
 def test_event_emitter():
     event_queue = EventQueue()
-    watch = ObservedWatch("/foobar", True)
+    watch = ObservedWatch("/foobar", recursive=True)
     event_emitter = EventEmitter(event_queue, watch, timeout=1)
     event_emitter.queue_event(FileModifiedEvent("/foobar/blah"))
 
 
 def test_event_dispatcher():
     event = FileModifiedEvent("/foobar")
-    watch = ObservedWatch("/path", True)
+    watch = ObservedWatch("/path", recursive=True)
 
     class TestableEventDispatcher(EventDispatcher):
         def dispatch_event(self, event, watch):
@@ -90,7 +89,7 @@ def test_observer_basic():
     observer = BaseObserver(EventEmitter)
     handler = LoggingEventHandler()
 
-    watch = observer.schedule(handler, "/foobar", True)
+    watch = observer.schedule(handler, "/foobar", recursive=True)
     observer.add_handler_for_watch(handler, watch)
     observer.add_handler_for_watch(handler, watch)
     observer.remove_handler_for_watch(handler, watch)
@@ -100,7 +99,7 @@ def test_observer_basic():
     with pytest.raises(KeyError):
         observer.unschedule(watch)
 
-    watch = observer.schedule(handler, "/foobar", True)
+    watch = observer.schedule(handler, "/foobar", recursive=True)
     observer.event_queue.put((FileModifiedEvent("/foobar"), watch))
     observer.start()
     time.sleep(1)